+++ src/sll/platform/windows.c
@@ -74,6 +74,22 @@ __SLL_FUNC void sll_platform_free_page(void* pg,sll_page_size_t sz){
 
 
 
+__SLL_FUNC __SLL_CHECK_OUTPUT sll_integer_t sll_platform_file_size(const char* fp){
+	HANDLE fh=CreateFileA(fp,GENERIC_READ,FILE_SHARE_READ|FILE_SHARE_WRITE,NULL,OPEN_EXISTING,FILE_ATTRIBUTE_NORMAL,NULL);
+	if (fh==INVALID_HANDLE_VALUE){
+		return 0;
+	}
+	LARGE_INTEGER sz;
+	if (!GetFileSizeEx(fh,&sz)){
+		CloseHandle(fh);
+		return 0;
+	}
+	CloseHandle(fh);
+	return sz.QuadPart;
+}
+
+
+
 __SLL_FUNC __SLL_CHECK_OUTPUT sll_time_t sll_platform_get_current_time(void){
 	FILETIME ft;
 	GetSystemTimeAsFileTime(&ft);